moved state to Content
