+++ src/App.module.css
@@ -2,3 +2,8 @@
   display: flex;
   flex-direction: row;
 }
+
+.App1 {
+  display: flex;
+  flex-direction: column;
+}
