+++ src/components/tabs/Home.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import style from '../../style/Home.module.css';
 
 class Home extends React.Component {
   // constructor(props) {
@@ -6,7 +7,7 @@ class Home extends React.Component {
   // }
 
   render() {
-    return <div>Home</div>;
+    return <div className={style.home_content}>Hello. It's home page.</div>;
   }
 }
