+++ src/components/tabs/Users/PopUp.jsx
@@ -1,10 +1,15 @@
 import React from 'react';
-import style from './PopUp.module.css';
+import style from '../../../style/PopUp.module.css';
 import InputText from '../../InputText';
 import DatePicker from 'react-date-picker';
 import namesOfInput from '../../../js/input_text';
 
 class PopUp extends React.Component {
+  addNewUser = () => {
+    this.props.close();
+    this.props.addNewUser();
+  };
+
   render() {
     const close = this.props.close;
     const state = this.props.getCurrentState();
@@ -32,8 +37,9 @@ class PopUp extends React.Component {
               onBlur={this.handleOnBlur}
               onChange={this.props.onChangeDirection}
             >
-              <option value='User'>User</option>
               <option value='Admin'>Admin</option>
+              <option value='Mentor'>Mentor</option>
+              <option value='Member'>Member</option>
             </select>
           </label>
           <label>
@@ -57,7 +63,7 @@ class PopUp extends React.Component {
             onChange={this.props.onChangeStartDate}
             value={this.props.state.startDate}
           />
-          <input type='button' onClick={this.props.addNewUser} value='Save' />
+          <input type='button' onClick={this.addNewUser} value='Save' />
         </form>
       </div>
     );
