--- src/components/tabs/Users/PopUp.module.css
@@ -1,41 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
-
-.modal {
-  font-size: 12px;
-  color: red;
-}
-.modal > .header {
-  width: 100%;
-  border-bottom: 1px solid gray;
-  font-size: 18px;
-  text-align: center;
-  padding: 5px;
-}
-.modal > .content {
-  width: 100%;
-  padding: 10px 5px;
-}
-.modal > .actions {
-  margin: auto;
-}
-.modal > .actions {
-  width: 100%;
-  padding: 10px 5px;
-  text-align: center;
-}
-.modal > .close {
-  cursor: pointer;
-  position: absolute;
-  display: block;
-  padding: 2px 5px;
-  line-height: 20px;
-  right: -10px;
-  top: -10px;
-  font-size: 24px;
-  background: #ffffff;
-  border-radius: 18px;
-  border: 1px solid #cfcece;
-}
