+++ src/style/Content.module.css
@@ -5,5 +5,5 @@
   height: 100%;
   color: aliceblue;
 
-  background-color: rgb(132, 153, 204);
+  /* background-color: rgb(132, 153, 204); */
 }
