+++ src/style/Error.module.css
@@ -1,0 +1,12 @@
+.content {
+  display: flex;
+  flex-direction: column;
+  width: 500px;
+  height: 70px;
+  margin: 100px;
+  padding: 10px;
+  color: rgb(0, 0, 0);
+  font-size: 20px;
+  border: 1px solid red;
+  border-radius: 5px;
+}
