+++ src/style/Home.module.css
@@ -1,0 +1,11 @@
+.home_content {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 380px;
+  color: black;
+
+  background-color: rgb(160, 52, 52);
+}
